Run Semantic Release in an environment with uv and npm installed
The Python CLI for Semantic Release provides the exact same functionality
as the GitHub Action. Running on the CI host with the CLI allows us to have
uv and npm available, so the Semantic Release build step to update lockfiles
is able to run. The `python-semantic-release version` command already has all
the defaults to run the build step, make a commit, push a tag, and push a
release (the same steps as before).

This uses `uvx` to ephimerally install Semantic Release, instead of adding it
as a proper dependency, since the version of Click required by Semantic Release
is old and unnecessarily constrains other packages. Developers are not expected
to run Semantic Release themselves locally.